@@ -1,21 +1,20 @@
 package com.bodega.api.io;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.Data;
-import org.hibernate.annotations.Type;
-
-import javax.persistence.*;
+import lombok.Getter;
+import lombok.Setter;
+import jakarta.persistence.*;
 import java.io.Serializable;
 import java.util.List;
 import java.util.UUID;
 
-@Data
+@Getter
+@Setter
 @Entity
 @Table(name = "categoria")
 @JsonIgnoreProperties({ "products" })
 public class CategoryEntity implements Serializable {
     @Id
-    @Type(type="org.hibernate.type.PostgresUUIDType")
     @Column(name = "id", columnDefinition = "uuid")
     private UUID id;
 
