@@ -1,31 +1,42 @@
 package com.bodega.api.io;
 
-import lombok.Data;
-import org.hibernate.annotations.Type;
+import lombok.Getter;
+import lombok.Setter;
 
-import javax.persistence.*;
-import java.util.Date;
+import jakarta.persistence.*;
+import java.io.Serializable;
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
 
-@Data
+@Getter
+@Setter
 @Entity
 @Table(name="venta")
-public class SaleEntity {
+public class SaleEntity implements Serializable {
     @Id
-    @Type(type="org.hibernate.type.PostgresUUIDType")
     private UUID id;
 
     @Column(name = "code")
     private String code;
 
-    @Column(name = "precio_venta") // todo in ms-product
+    @Column(name = "precio_venta")
     private int salePrice;
 
-    @Column(name = "fecha_venta")  // todo in ms-product
-    private Date dateRegister;
+    @Column(name = "fecha_venta")
+    private LocalDateTime dateRegister;
 
-    @OneToMany(cascade = CascadeType.ALL, mappedBy = "sale")
-    private List<SaleDetailEntity> saleDetails;
+    @Column(name = "paypal_id")
+    private String paypalId;
+
+    @OneToMany(cascade = CascadeType.ALL, mappedBy = "sale",  fetch = FetchType.EAGER)
+    private List<SaleDetailEntity> saleDetail;
+
+    @Column(name = "usuario_id", columnDefinition = "uuid")
+    private UUID userId;
+
+    @ManyToOne
+    @JoinColumn(name = "usuario_id", referencedColumnName = "id", insertable = false, updatable = false)
+    private UserEntity user;
 
 }
