@@ -1,26 +1,26 @@
 package com.bodega.api.io;
 
-import lombok.Data;
-import org.hibernate.annotations.Type;
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.Setter;
 
-import javax.persistence.*;
 import java.io.Serializable;
 import java.util.List;
 import java.util.UUID;
 
-@Data
+@Getter
+@Setter
 @Entity
 @Table(name = "producto")
 public class ProductEntity implements Serializable {
     @Id
-    @Type(type="org.hibernate.type.PostgresUUIDType")
     public UUID id;
 
     @Column(name = "nombre")
     private String name;
 
     @Column(name = "precio_unitario")
-    private int unitPrice;
+    private double unitPrice;
 
     @Column(name = "cantidad")
     private int quantity;
@@ -28,7 +28,7 @@ public class ProductEntity implements Serializable {
     @Column(name = "imagen")
     private String image;
 
-    @Type(type="org.hibernate.type.PostgresUUIDType")
+
     @Column(name = "categoria_id", columnDefinition = "uuid")
     private UUID categoryId;
 
