@@ -4,13 +4,14 @@ import com.bodega.api.io.ProductEntity;
 import com.bodega.api.repository.ProductRepository;
 import com.bodega.api.service.ProductService;
 import com.bodega.api.shared.dto.ProductDto;
-import com.bodega.api.ui.model.request.ProductRequest;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
+
+import java.util.UUID;
 
 @Slf4j
 @Service
@@ -22,7 +23,16 @@ public class ProductServiceImpl implements ProductService {
     @Autowired
     ProductRepository repository;
 
-    @Override
+  @Override
+  public Mono<ProductDto> getProduct(UUID id) {
+    var product = repository.findById(id);
+    if (product.isEmpty()) {
+      return Mono.error(new RuntimeException("No existe el product con id: " + id));
+    }
+    return Mono.just(mapper.map(product.get(), ProductDto.class));
+  }
+
+  @Override
     public Flux<ProductDto> getProducts() {
         return Flux.fromIterable(repository.findAll())
                 .map(productEntity -> mapper.map(productEntity, ProductDto.class))
@@ -30,8 +40,8 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public Mono<ProductDto> createProduct(ProductRequest createProductRequest) {
-        return Mono.just(mapper.map(createProductRequest, ProductEntity.class))
+    public Mono<ProductDto> saveProduct(ProductDto productDto) {
+        return Mono.just(mapper.map(productDto, ProductEntity.class))
                 .flatMap(productEntity -> {
                     ProductEntity createdProduct = repository.save(productEntity);
                     return Mono.just(mapper.map(createdProduct, ProductDto.class));
